# Compare performance with flatrtree