--- a/go.mod
+++ b/go.mod
@@ -2,14 +2,24 @@ module github.com/tidwall/sick-spatial-join
 
 go 1.17
 
+require github.com/tidwall/geojson v1.3.4
+
 require (
-	github.com/tidwall/geojson v1.3.4
-	github.com/tidwall/rtree v1.3.1
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/furstenheim/nth_element v0.0.0-20190419153334-6ae3dd61145d // indirect
+	github.com/invisiblefunnel/flatqueue-go/v2 v2.0.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.8.1 // indirect
+	github.com/tidwall/rtree v1.3.1 // indirect
+	golang.org/x/exp v0.0.0-20220602145555-4a0574d9293f // indirect
+	google.golang.org/protobuf v1.28.1 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 require (
+	github.com/flatrtree/flatrtree-go v1.0.0-rc1
 	github.com/tidwall/geoindex v1.4.4 // indirect
-	github.com/tidwall/gjson v1.12.1 // indirect
+	github.com/tidwall/gjson v1.12.1
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.0 // indirect
 	github.com/tidwall/sjson v1.2.4 // indirect