```console
% git status
On branch master
Your branch is up to date with 'origin/master'.

nothing to commit, working tree clean
% rm data/output.csv   
% go run main.go       
Loading neighborhoods... 0.03 secs
Loading violations... 2.61 secs
Joining neighborhoods and violations... 5.04 secs
Writing output... 0.37 secs
Total execution time... 8.03 secs
% wc -l data/output.csv 
 8652609 data/output.csv
% head -n5 data/output.csv
anon_ticket_number,neighborhood
1777797,Center City East
1777798,Chinatown
1777799,Wister
1777801,Center City East
% rm data/output.csv 
% git checkout flatrtree
Switched to branch 'flatrtree'
Your branch is up to date with 'origin/flatrtree'.
% go run main.go          
Loading neighborhoods... 0.03 secs
Loading violations... 2.58 secs
Joining neighborhoods and violations... 4.05 secs
Writing output... 0.39 secs
Total execution time... 7.05 secs
% wc -l data/output.csv   
 8652609 data/output.csv
% head -n5 data/output.csv
anon_ticket_number,neighborhood
1777797,Center City East
1777798,Chinatown
1777799,Wister
1777801,Center City East
```